requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sse-starlette==1.6.5
plexapi==4.15.4
python-dotenv==1.0.0